FIX: ненужное удаление отсутствующей папки при очистке в начале обмена

+++ Shop/Structure/Service/Load1CV3/ExchangeUtil.php
@@ -39,7 +39,7 @@ class ExchangeUtil
                 closedir($dh);
             }
             rmdir($path);
-        } else {
+        } elseif (file_exists($path)) {
             unlink($path);
         }
     }
